Add CLI command for new start flow
